requirements.txt
faiss-cpu>=1.7.4
numpy>=1.24.0
sqlalchemy>=2.0.0
aiosqlite>=0.19.0
aiofiles>=23.0.0
pypdf>=3.0.0
markitdown>=0.0.1a2
openpyxl>=3.1.0
rank-bm25>=0.2.2
jieba>=0.42.1
httpx>=0.27.0
starlette>=0.27.0
uvicorn>=0.23.0
pdf2image>=1.16.0